requirements.txt
requests
pandas
psycopg2
nfldb
nflgame-redux